relnotes: Compare first alpha against latest vX.Y.0.
Previously, the script chose v1.5.0-alpha.2 as the diff base for
v1.6.0-alpha.1. Now it will choose v1.5.0 in that case, so we won't list
changes again that already made it into the previous release.

I decided not to use the latest v1.5.x as the diff base, because
technically those should only contain backports of work that was
actually done against the 1.6 branch.