git checkout -B resets existing branches.  Check if branch exists.
And highly suspect in damaging the master branch state (and tags) is
the newish fetch/rebase action in gitlib::git_push_access().
With the recent introduction of gitlib::is_repo_admin(),
gitlib::git_push_access() is obsolete.

Further, refactor the way tags are applied in prepare_tree() for the
master/alpha case in order to ensure we don't lose the tag on a rebase.