Assign specfic members to the publishing bot issue
The members to assign to the publishing bot issue are the members of the
github team 'kubernetes/publishing-bot-reviewers'. We query the API to
get all the members and assign to them individually, because GitHub
cannot assign to teams.

However, github can reqeust reviews ("/cc ...") from teams, so we also
do that.