anago: direct echo command for easier copy paste
When finishing a successful gcbmgr stage, anago helpfully emits
sample gcbmgr and anago command for the next step of release, but
logecho use of fmtlen can cause the long "--buildversion" to be on
a separate line.  This changes the echo so copy paste reuse of the
emitted command is easier.

Signed-off-by: Tim Pepper <tpepper@vmware.com>