debian: Restructure directory
Renames the bionic/ dir to packages/ and removes the symlinks for
individual Debian/Ubuntu releases. We've been building with the same
package templates for some time now, which suggests that the symlinks
are extraneous. Additionally, we remove symlinks in the kubeadm
package and bring config files for each package up a few levels.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>