gitlib: allow using ssh remote url
The remote url when using:

- https: https://github.com/kubernetes/release.git
- ssh: git@github.com:kubernetes/release.git

For ssh, there is no forward slash before `kubernetes`. So remove it
from `expectedRemote` so that it can match both https and ssh.

If string constants are used for regular expressions with awk, dynamic
regex is used and the string is scanned twice.
Ref: https://www.gnu.org/software/gawk/manual/html_node/Computed-Regexps.html

So add extra back slashes to actually escape the backslash for matching
the bracket in "fetch", and don't escape the forward slash as it isn't
needed (only back slashes need to be escaped).