[gcbmgr] Fallback to master's version of the kubecross image
One of the first steps in GCB is to compile go things, e.g. the
`blocking-testgrid-tests` binary. Theses binaries are to be compiled
with the same go version as the version of kubernetes itself. Therefore
we use the kubecross image as the image for this GCB task.

Now, when the branch og k/k does not exist, that fails. This can be the
case of an beta cut, because the branch is only to be created by anago.
So for these cases we allow a fallback to the master branch. It is
clearly logged however, that we couldn't determine the kubecross image
version to use and did the fallback.

Also note: This is only for building the binaries used in the tests, not
for building kubernetes itself.