Publishing md5s / sha1s through 1.16, put everything into shaNNNsums
This script appears to be used for previous versions, at least in CI,
so we should keep publishing sha1 & md5 files.

Also put all our assets into the SHA256SUMS / SHA512SUMS files.  This
means that we'll end up signing everything we publish (at little to no
cost).  It also means that SHA256SUMS and SHA512SUMS files can be used
for mirroring.  Also, we can detect _extra_ files, which is also
potentially a problem for release attestation.  The implementation
runs the hash twice, but the code is much simpler for it.

We also unroll the loop for clarity, and so we can easily
version-guard md5/sha1 as part of a future deprecation.