build/debs: Switch CI version marker from "latest" to "k8s-master"
Changing this version marker due to an idiosyncrasy with what we
consider latest. We run two push-build jobs (ci-kubernetes-build
and ci-kubernetes-build-fast) which push the latest.txt version
marker. The fast job only pushes a linux-amd64 build of Kubernetes
which means if someone tries to build debs for the latest CI version
for all architectures, it will sometimes fail (depending on whether
the latest marker is pointing at a cross-build or not). The k8s-master
version marker is always a cross-build, so we use it here instead.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>