[testgridshot] Use the upload key when provided or bail out otherwise
On 2019-09-03 the image hosting service we use disabled anonymous
uploads. Therefore, if you want to use this script, you need to create
an account at vgy.me and a user key.

This makes the script pretty useless. Maybe we find a way to upload
images to GH's CDN, but for now there does not seem to be a propper API
and only workarounds[1] for that available.

[1]: https://gist.github.com/vinkla/dca76249ba6b73c5dd66a4e986df4c8d