gcb/build: Update substitutions and disable image pushes on fast builds
Here we remove the default substitutions for _BUCKET and _REGISTRY,
which forces a user to explicitly define them within variants.yaml or
from the command line.

We also remove the REGISTRY field from the `build-ci-fast` variant,
which will cause resulting images from a fast build not to be pushed to
GCR. Images should only be pushed from a cross build.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>