anago: Ensure kube-cross:local matches the project we stage against
This tweaks the configuration of local_kube_cross so that it builds
an image within the staging project. We're changing this as the current
GCB stage build config points to k8s.gcr.io, which we do not have access
to push to from the staging GCP project (k8s-staging-release-test).

Signed-off-by: Stephen Augustus <saugustus@vmware.com>