lib/release: Set mock bucket and remove Google-isms in bucket/registry
The current logic for setting globals on buckets and registry names is
based on a local variable (gcb_user=gcb@google.com), which causes the
bucket names to always be appended with "-gcb". Here, we clean up some
of the logic to instead determine bucket/registry names based on global
variables in the release library.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>