branchff: Remove hack/update-openapi-spec.sh run from tool
This is an optimization, which also allows us to perform a true branch
fast-forward using "branchff".

Prior to this commit, we were running hack/update-openapi-spec.sh,
which:
- installs a bunch of prerequisites, including etcd
- builds the apiserver
- runs an apiserver to retrieve the OpenAPI spec
- rewrites a version field within the resulting swagger.json to match
   the targeted release branch

And then we commit the result before pushing to the upstream release
branch.

This is obviously wasteful and runs contrary to the definition of a
fast-forward.

A fix to "null" the version field in the swagger.json has been
submitted, so this commit removes that script logic to match.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>