Use a github wiki to store the images in
Instead of pushing the images to a image hosting service we now add them
to the git repository of the wiki for a GitHub repository (`k/release`
by default). For that, specifically for the `git push` to work, users
need to provide a GitHub token in the env variable `$GITHUB_TOKEN`.