Add make targets to update dependencies
For now, the go target is the most important one, we don't really have
other, non-go, dependencies for now (except the bazel stuff).

For go (`make update-go-deps`) we
- pull in all updated deps, also for testing modules
- tidy up the go modules
- verify all go module dependencies
- run the tests
- update the bazel stuff

Generally for updating dependencies (`make update-deps`) we ask the user
to PR those changes and present them with a short summary on what has
changed. We do not create a commit or stage anything, we leave that to
the user.

We set the new targets to be silent and use a target-specific variable
to force the usage of go modules.