Switch to python3
Why?
```
DEPRECATION: Python 2.7 will reach the end of its life on January 1st,
2020. Please upgrade your Python as Python 2.7 won't be maintained after
that date. A future version of pip will drop support for Python 2.7.
More details about Python 2 support in pip, can be found at
https://pip.pypa.io/en/latest/development/release-process/#python-2-support
```

One of the images this image is based on installs mercurial via apt.
This package has a hard dependency on python2. To get rid of python2 and
install python3 we just unsinstall every python2 package for now. This
means we also uninstall mercurial.
In case this breaks something, we can still bring in mercurial via pip3.

In the long run, we should work with upstream, debian in our case, to
get rid of mercurial's the hard dependency on python2. A recent enough
version of mercurial works with python3, so technically it should be
possible to drop that dependency for the debian package.