Bind git branch verification to `Repo`
This is the first part of the git refactoring. The verification if the
remote branch exists will now be done via the `Repo`, which will be
initialized by the newly added `--repo` command line option.

SSH authentication has been added on top, whereas the other functions
will be migrated in follow-up PRs.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>