Do not close stderr when we want to hide the error
'2>&-' closes standard error. Some tools in the current version do not
like that anymore (gsutil for example). Therefore we change the
behaviour to keep stderr open, but redirect to /dev/null.

The result should be the same: for certain commands we hide any error.
Therefore we opportunistically change this for all commands and not only
for gsutil.