Add git package unit tests
This commit adds all possible offline unit tests to the git package.
This excludes methods which are dedicated to GitHub repository
creation/update.

The command package had to be updated to contain an additional command
creation function which is capable of specifying the working directory
where the command should be run. Two convenient methods `RunSuccess()`
and `RunSilentSuccess()` have been added as well to reduce the status
checking overhead it not necessary. Tests are now available for these
new methods too.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>