[relnotes] Introduce a notes.Client to be able to mock the interaction with GitHub
We created an interface type that holds all the interactions needed with
github. We have one concrete implemenation which is just a thin wrapper
around a regular old *github.Client.

This should allow us to mock any GitHub interaction and run some tests
in isolation.
Evyerthing is setup to have the mock automatically created or updated by
a simple `go generate ./...`.