[relnotes] Don't compute the same things over and over
This pulls out the REs to check the PR bodies against from the nested
loop. There is no reason to set them up and compile the REs over and
over again, they don't depend on any data only available inside the
loops.
This also makes the code quite a bit easier to read.