[relnotes] Remove GitHubAPIOption due to lack of value
The GitHubAPIOption does not give us much value, it makes everything a
bit more complicated. IMHO, it does not make much sense to pass those
options around and compute the "real" options just before we use them.
The way as the options are implemented here do not guard against
changing them midway execution and don't make any calls any more
readable.
If we'd want to, we could implement the functional options pattern for
the Gatherer.