kubepkg: Add getBuildArch function and rename cfg to buildConfig
Now that we're building for multiple distro types, the conditional logic
for translating goArch to a build arch no longer works. Here we create
buildArchMap (map[string][BuildType][string]) to represent the various
GoArch/BuildArch translations, remove the conditional logic for this in
buildPackage() and use a simple function to retrieve values from the map.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>