kubepkg: Support templating as a standalone step
Here we add flags (--template-dir, --spec-only) and make some minor
changes to the input/output directories to allow a user to just create
specs instead of also creating packages. This was more of an
intermediate step before, which meant a user doesn't have the
opportunity to inspect the specs before building the packages.

The immediate useful case for this is that we can now commit specs to
the repo. Committed specs would allow us to reproduce packages at points
in time, like releases.

Example: specs/1.17/1.17.1/packages/kubelet/<arch>/kubelet.spec

Signed-off-by: Stephen Augustus <saugustus@vmware.com>