kubepkg: Fix packageDefinition pointer copy
In a previous commit we started passing pointers of PackageDefinition in
various functions. While this is mostly fine for referencing values, in
buildPackage(), we're actually assigning values to underlying struct.

As we loop over this data a few times, we're essentially poisoning each
build run with data from a previous one.

To fix this, we now create a new PackageDefinition struct, copy the
underlying data from the reference PackageDefinition (*ptr2 = *ptr1), and
then only modify the copy.

This was noticed during runs where the channel selections included the
nightly channel. Nightly runs pull a CI build version of Kubernetes
(vx.y.z-{alpha,beta,rc}.m+n):
- x: major
- y: minor
- z: patch
- m: build number
- n: commit-ish

We do replaces on "+" to "-", so that the Kubernetes version is valid to
use for packaging, but that's only done _after_ constructing a download
link (which needs to include the "+" to be valid). So the first nightly
package build would succeed and then subsequent builds would fail with
an invalid URL.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>