Fix minor-to-latest discovery mode
The minor to latest discovery mode had no usage because it was not
assuming the correct start revision. From a process perspective, we have
to use the first RC (code freeze) because this is the last time that the
branch got synced with the master.

This is the same commit as the merge base of the previous release
branch, which we now use to provide a more clean technical solution.

We also rename the discovery mode to `mergebase-to-latest` to reflect
the new behavior.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>