Add changelog remote lookup feature
We now correctly assume that a non-patch final version needs to have a
remote changelog available and failure if this one does not exist.

We also have to provide the target tag to the changelog subcommand to be
able to find the right release notes generation approach.

The git package has been adapted to contain a more robust way to
retrieve tags for a dedicated branch.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>