kubepkg: Refactor directory structure
Here we move files from the build/* directories into the cmd/kubepkg
directory. Eventually, this will enable us to run GCB with a smaller
build context, but that's predicated on deciding on a versioning scheme
for k/release and potentially creating a kubepkg submodule.

In the interim, take note that the GCB config and Dockerfiles for
kubepkg are in the root of the repo, to allow the go.mod to be
included in the Docker build context.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>