kubepkg: Support container builds and GCB builds
Tweak the Dockerfiles for kubepkg to reference the new asset locations
and update the GCB config to build the container images for debs and
rpms.

As the current tool we use to build rpms (rpmbuild) has requirements
that need to be built on Fedora-like systems, we're building a separate
container image for the "kubepkg rpms" command usage.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>