anago: Remove OpenAPI spec version bump logic
We recently set the OpenAPI spec "version" field to "unversioned" in
k/k(84654) (on master branch ahead of 1.18, which was also cherry picked
also to release branches 1.15, 1.16, and 1.17), which means we no longer
need to handle rev'ing the OpenAPI spec within anago.

This removes the code paths and function that does that.
This also means anago's git_tag() function is only called in one place
and can be inlined.

That had become a highly overloaded function, but it is down to a mere
two or four flavors depending on the stateful nature of its usage.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>
Co-authored-by: Tim Pepper <tpepper@vmware.com>