anago: Ensure ordered releases by using a sorted release key array
Depending on the type of Kubernetes release we run with anago, we may be
cutting multiple releases and creating multiple tags on GitHub.

For point zero and patch releases (vx.y.z), we first run an "official"
release, followed by the next beta for that release branch
(vx.y.(z+1)-beta.0).

For the first beta of a cycle (vx.y.z-beta.0), we cut a release branch
(release-x.y) and push the next alpha tag to the master branch
(vx.(y+1).z-alpha.1).

For these activities to happen correctly, the process must run in the
correct order: official --> rc --> beta --> alpha

Prior to this commit, in several steps, we were iterating over the
RELEASE_VERSION associative array with the syntax
${!RELEASE_VERSION[@]}. Iterating over an associative array does not
guarantee ordering.

To fix this, we create a new non-associative array,
ORDERED_RELEASE_KEYS, which holds the ordered keys of RELEASE_VERSION.

From there, we then refactor each of the iterators using
${!RELEASE_VERSION[@]} to instead use ${ORDERED_RELEASE_KEYS[@]} and use
the keys therein to access the underlying values in RELEASE_VERSION.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>