[builder-img] Add multiarch support
The k8s-cloud-builder image is based on the cross build image, which
already brings multiarch support by installing cross build toolchains
for the supported architectures.

Before this commit we were cleaning up the image pretty rigorously,
which specifically means we also removed the gcc and the cross build
tool chains.

Now we soften the cleanup a bit and keep the cross build toolchains in
the main image. We don't really need it for the release process
(gcb/{stage,release}/...), but is allows us to run CI builds / `make
release-in-container` (gcb/build/...) without the need to maintain
another image.