anago: ensure beta tag is on unique commit
We have problems where 'git describe' on a commit with multiple tags does
not emit the tag we might logically prefer.  For example, if you are
checked out at:

	commit e7f962ba86f4ce7033828210ca3556393c377bcc (tag: v1.16.6-beta.0, tag: v1.16.5)
	Merge: 7d00adae4c2 67d2a3438f0
	Author: Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com>
	Date:   Fri Jan 10 17:41:37 2020 -0800

and run git describe you (may) get the beta tag:

	$ git describe
	v1.16.6-beta.0

We have code that requires that output be "v1.16.5", not the beta tag.
If nothing else the output of git describe ends up in the GitVersion
string emitted by our binaries (eg: 'kubectl version').  Its obviously
unacceptably confusing if the v1.16.5 binary reports that it is
v1.16.6-beta.0.

And apparently there are workflows where these beta tags are used instead
of the prior official (or alpha) tag, so we cant simply drop the beta
tagging lest we cascade yet more failure.

The easiest way short term to ensure this spread of tags versus commits
then is to insert a blank commit before any beta tagging.

Signed-off-by: Tim Pepper <tpepper@vmware.com>