Repo update should persist repo state
The update of a repository will cause the issue that all local changes
will be overwritten. This is not a desired behavior and we will now
update the repo with a usual "pull --rebase --autostash". This also
changes the way how we have to checkout branches because the go-git
implementation will now work any more with that new strategy. For now we
use the simple "git checkout" shell-out command which seems to have a
more robust implementation.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>