Add CHANGELOG_DIR variable to set authoritative changelog location
Recent changes to the changelog location in kubernetes/kubernetes mean
we need to ensure all places that reference the changelog location can
be changed on the fly AND consistently.

Adding this because stages on a to-be created branch fail on the removal
of the previous changelogs. This code branch was using a wildcard, so we
didn't pick it up in our initial search/replace.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>