gcb: Hardcoded k8s-cloud-builder image location in configs
Without this, the stage/release will expect a corresponding
k8s-cloud-builder image in the kubernetes-release-test and
kubernetes-release GCP projects. This was originally using $PROJECT_ID
to allow some configuration flexibility when testing images, but as the
secrets are hardcoded in our templates as well, they're already
essentially only usable for a single GCP project.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>