Make PackagesAvailable() more universal
The utility function `PackagesAvailable()` now should work more
universally and independently from the host OS by just checking if a
matching subset of commands is available, like `dnf` and `rpm`.

This means we do not have to rely on the host OS any more. Testing has
been adapted as well.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>