WIP: Pass a list of NoteCategory to template instead of map of kind to noteslice
Introduces a NoteCollection type that can be passed to the template. This is
need so that the kinds can be rendered in priority order (PR#1148). In the
current implementation the template cannot render the notes in order because
the NotesByKind map is unordered.

If we like this idea I can fully implement it.