Make git.GetRepoURL infallible
The error due to parsing the `DefaultGithubURLBase` is technically not
possible and will fail every time if `DefaultGithubURLBase` is wrong. To
avoid that, we now manually create the always correct URL and let
`GetRepoURL` only return the resulting `string` type. The tests and
dependent functions have been adapted as well.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>