lib/release: Add image validation
Here we add 'release::docker::validate_remote_manifests()', a function
which validates the existence of image manifests on a remote registry
by doing the following:
- Using 'skopeo inspect' to retrieve an image manifest
- Looping through the architectures and checking for an
  architecture-specific digest within the manifest using jq

We do some minor cleanup by using 'find' and 'basename' to populate
arrays for 'for' loops.

Additionally, we validate the correct amount of arguments for
'release::docker::validate_remote_manifests()' and
'release::docker::release()' using 'common::argc_validate'.

This function closely resembles 'release::docker::release()' and could
probably be consolidated. We make the decision to not do this refactor
here to minimize the changes to surrounding functions.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>