[VDF] Partial revert of k8s.gcr.io cutover
This partially reverts commits:
- 29b8e1da5c4bc08d1bbfc64b61e68e59bbe4f70b
- 4343e86a4fb406523445eed3173334633d6ef801
- dcaf5c063cd8929834fe3916d2fb5408b7954149

From Linus:
"We found a hard-coded internal reference that we had missed which
caused some large fraction of traffic from certain regions to be sent to
the wrong place. Because it is Friday afternoon, we will halt the
rollout, do a fast rollback, and try again early next week. Sorry!"

Signed-off-by: Stephen Augustus <saugustus@vmware.com>