kube-cross: Add k8s-cloud-builder functionality
Currently the k8s-cloud-builder image depends on the kube-cross image
version for staging/releasing Kubernetes. In our current state, we'd
require a bump to the kube-cross version if the content of
k8s-cloud-builder changes. This is both counter-intuitive and
time-consuming from a promotion perspective.

This commit moves the k8s-cloud-builder Dockerfile content into
kube-cross allowing us to maintain a single image for
building/staging/releasing Kubernetes.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>