Better way to override local directory for artifacts
The current override is not pleasant to use, let's keep it simple with
a single env var `KUBE_LOCAL_ARTIFACTS` which is the directory to where
the binaries can be found. No need to muck around with `go env` or use
the hard coded path off k8s.io/kubernetes that is specific to one of the
two build systems in kubernetes.

This way it does not matter how the binaries were built, just that they
are available in a well-known tree structure. We use the same tree
structure in the url for the downloads from gcs as well as a bonus.

Signed-off-by: Davanum Srinivas <davanum@gmail.com>