Add GetJobCache function from releaselib.sh
Due to the migration of the set_build_version function from the release
lib we now convert the get_job_cache function to a golang based
implementation. The signature of the function has changed that we do not
write the results into a directory any more and directly return a new
tpye `JobCache`, which is a dictionary for build numbers and versions.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>