Remove relnotes script
We now remove the bash relnotes script and re-create a `pkg/announce`
package which will later on create the release and branch announcement
for krel/anago. Code parts which are not needed any more in `pkg/patch`
have been cleaned-up as well to be able to remove more bash parts.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>