Unlist release notes during processing
Some release notes look like this:

````
```release-note
This is a release note
which fills multiple
lines in the block.
```
````

whereas others look like this:

````
```release-note
- This is a release note
  which fills multiple
  lines in the block.
```
````

After this patch both will look the same after processing. This provides
the benefit for a more uniform formatting for the website as well as a
better default markdown formatting.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>