anago: Push git objects after GCS/GCR artifacts
Pushing artifacts to GitHub signifies for a lot of consumers, especially
those that watch kubernetes/kubernetes releases that a release is
complete.

We've had recent instances where there were intermittent issues with
GCS/GCR artifacts which required us to workaround the GitHub tag checks,
so this should partially mitigate that.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>