Remove find_green_build and all of its dependencies
The questionable logic behind "finding a green build" for a release has
not been used by us since a long time. We usually pre-check the CI
before going to cut a release, which means that we can finally stick to
the ci/latest Jenkins build version on the corresponding branch.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>