Remove prin and sample script
We do not want to add more bash based logic to the repository which is
relevant around cutting releases. Therefore we remove the
`script-template`. The `prin` script seems unused by release folks as well,
so this has to go too to be able to remove some more functions in `lib`
later on.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>