Split Go test suite unit and integration
We decided that all the tests part of the `./cmd` package are
integration tests, and they should run separately from the other
packages.

This commit create a new `make test-go-unit` that runs the unit tests.
It runs as part of the old `make test` target.

There is a new target called `make test-go-integration` that runs the
`./cmd` tests only.