Remove inferable flags from `krel anago push`
Since `krel anago push` is now in place I did a comparison of the flag
values passed to the subcommand. Some of the flags can be inferred by
the `--build-dir`, whereas others are not required at all. With this
patch, we remove the flags

- `--local-gcs-stage-path`
- `--local-release-images-path`
- `--gcs-suffix`
- `--staged-bucket`

and infer their values from the other provided flags.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>