Always validate images against the provided registry
The logic to flip the container registry for the image validation was an
artifact from old times, where we had the promotion in a different
place. We can later on decide if we still want to verify the images on
`k8s.gcr.io` in a dedicated verification step on `release`, not on
`stage`.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>