Delete BUILD_AT_HEAD key in krel gcbmgr if not set
Before this patch we set the GCB substitution to `""`, which caused a
failure like:

```
ERROR: (gcloud.builds.submit) INVALID_ARGUMENT:
generic::invalid_argument: key in the template "_BUILD_AT_HEAD" is not
matched in the substitution data;
```

We now delete the key in the map if no build version is provided. The
issue was only reproducible on macOS, not on Linux.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>