Integrate workspace preparation into `krel release`
We now integrate the workspace preparation into krel release. For that
we add a new `Options` API `Bucket()`, which can be used to retrieve the
right target bucket based on the value of `NoMock`. We also move the
options into the stage/release implementations to be able to access them.

Mocks and tests have been updated as well.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>