Add release version and artifacts push to krel stage/release
We now add the artifacts pushing and generation of the release version
to `krel stage/release`. This does right now not work for releases where
we create a release branch, because this variable will be set by
`SetBuildCandidate`, which is not implemented yet.

It does also not work stage, because the release build is not
implemented there.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>