Do not escape `#` in notes any more
We attempted to fix the direct usage of code blocks inside release notes
with the escaping of `#`.

See https://github.com/kubernetes/release/issues/681

This scenario is not supported by us, because If people want to use code
blocks within the release notes, they usually have to define a new one
in the markdown, like this:

`````
```release-note
````
> echo "Hello world"
````
```
`````

This means we do not really have to escape the `#`-values, because this
will result in breaking HTML anchored links.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>