Add gcb package
The subcommand `krel gcbmgr` should move to a plain execution of `krel
stage/release` in the future. We still need the gcbmgr logic to achieve
that. To be able to use that logic directly in `krel stage/release`, we
must have it available as a package outside of the `cmd` package. This
is mainly what this commit achieves.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>