pkg/build: Deprecate the GCSSuffix option and introduce GCSRoot option
Testing with the `--gcs-suffix` flag has been frustrating.

It's possible the original shell library never properly handled this
option and it was never noticed because the option isn't exercised in
our releases or CI jobs.

Here we replace its usage with `--gcs-root`.

If specified, it will override BuildType.

When unset:
  - BuildType: "ci"
  - final path: gs://<bucket>/ci

When set:
  - BuildType: "ci"
  - GCSRoot: "new-root"
  - final path: gs://<bucket>/new-root

Signed-off-by: Stephen Augustus <saugustus@vmware.com>