pkg/build: Fully remove GCSSuffix option
`build.Option.GCSSuffix` was deprecated in a previous PR, pending review
of the codebase for usage.

An update to the `ci-kubernetes-build-no-bootstrap` removes it reliance
on the flag.

Here's a look through the codebase:
- https://cs.k8s.io/?q=gcs-suffix&i=nope&files=&repos=
- https://cs.k8s.io/?q=gcsSuffix&i=nope&files=&repos=
- https://cs.k8s.io/?q=KUBE_GCS_RELEASE_SUFFIX&i=nope&files=&repos=

All references are plumbing, not actual usage, and will be removed in
follow-ups to k/test-infra and kubetest.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>