Disable noClobber for uploading source tree tarball
This commit disables noClobber flag for gsutil cp when uploading the
source tree tarball to GCS. When this flag is disabled, it allows
gsutil to overwrite the file if it already exists.

This is useful if we're staging multiple releases and all releases have
the same buildversion (e.g. when testing staging).

Signed-off-by: Marko Mudrinić <mudrinic.mare@gmail.com>