git: rev-parse with specified tag regexp
The current implementation of `RevParse` tries to match the incoming
revision with `v\d+\.\d+\.\d+.*`. This works in most cases. But this
may not always be the case for downstreams(as a user of relesae-notes
tool).

This change introduces a new flag `--tag-regexp` so that the way to
match git tag names can be customized.