Binary package: Base module for binary analysis
This commit initializes the new binary package which will
house our code for performing verification and analysis of
binary executables.

This first iteration adds the base implementations to open
and extract data from the executable headers of ELF (Linux),
Mach-O (MacOS) and PE (Win) binaries.

The binary object has two methods:

* `Arch()` returns the GOARCH label obtained from analyzing a binary
* `OS()` returns the GOOS label that indicates the operating system of the binary

The `binary` object has a Header object which can be queried to
get some platform sepcifics info. Has currently one method: `WordLength()`
which returns an integer (32/64) which will determine if the file is a
32 or 64 bit binary.i

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@uservers.net>