Fix docker buildx image builds
It seems that we broke the qemu initialization for applying
multi-architecture builds via docker buildx. We now use the latest
`multiarch/qemu-user-static` image to initialize docker buildx and
register a new builder which should support all required architectures.

This means that we also can remove the vendored `register.sh` script
from this repository because it is not being used any more.

Beside that change, we also have to `--load` the images built via
`docker buildx`, otherwise they will remain within the local build cache
and will not being populated into the image store.

Signed-off-by: Sascha Grunert <mail@saschagrunert.de>